pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enterprise-mcp-server"
version = "2.0.0"
description = "Enterprise MCP Server for Claude — Jira, GitHub, Confluence, Slack, PagerDuty, Datadog (38 tools)"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [
    {name = "Adrien Thomas", email = "adrien@example.com"},
]
keywords = ["mcp", "claude", "anthropic", "enterprise", "jira", "github", "confluence", "slack", "pagerduty", "datadog"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp>=1.0.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.3",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "redis[hiredis]>=5.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "cryptography>=42.0.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.2",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "bandit>=1.7.7",
    "types-redis>=4.6.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
enterprise-mcp = "enterprise_mcp.server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.hatch.build.targets.wheel]
packages = ["src/enterprise_mcp"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
